Add files via upload
The Excel workbook 1 explores the following skills:

First Problem:
a. Creates a scatterplot and answers if the relationship is a linear one. 
b. Superimpose a trend line on the scatterplot and use the option to display the equation and the R2 value. Predict the final score.
c. Run a regression analysis to confirm the trend-line equation from part b. Then, define what the standard error of estimate says about the accuracy of the prediction requested in part b.

Second Problem:
a. Estimates a multiple regression equation.
b. Interprets each of the estimated regression coefficients. 
c.  Interprets the s standard error of estimate se and R2 for these data.

Third Problem:
a. Estimates an appropriate multiple regression equation to predict the cost of shipping a given package. 
b. Interprets the estimated regression coefficients. 
c. Answers which cargo type is the costliest to ship according to the estimated regression equation.
d. Answers how well the estimated equation fits the given sample data and how the model’s goodness of fit could be improved.
e. Predicts the cost of shipping a package with semi-fragile cargo.